Add files via upload
Fedora 43 installation walk through guide, which installs all of the needed Fedora packages, Intel oneAPI Base Toolkit, and llama.cpp inferencing tool, which is a great tool to run some LLM's.  There is even some shell script examples at the bottom of this text document, which runs in a server setting, or a cli (terminal) setting.